refactoring(explorer): adde PatternBase class
